requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.1
numpy==1.14.0
PyMySQL==0.8.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-telegram-bot==9.0.0
pytz==2017.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.1
urllib3==1.22
Werkzeug==0.14.1
